Updated the Models to hold a fieldId and url for the images, Videos and Mades changes to COntrollers Accordingly.

+++ src/utils/cloudinary.js
@@ -1,33 +1,39 @@
-import {v2 as cloudinary} from "cloudinary"
-import fs from "fs"
+import { v2 as cloudinary } from "cloudinary";
+import fs from "fs";
 
-
-cloudinary.config({ 
-  cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
-  api_key: process.env.CLOUDINARY_API_KEY, 
-  api_secret: process.env.CLOUDINARY_API_SECRET 
+cloudinary.config({
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+  api_key: process.env.CLOUDINARY_API_KEY,
+  api_secret: process.env.CLOUDINARY_API_SECRET,
 });
 
-const uploadOnCloudinary = async (localFilePath) => {
-    try {
-        if (!localFilePath) return null
-        //upload the file on cloudinary
-        const response = await cloudinary.uploader.upload(localFilePath, {
-            resource_type: "auto"
-        })
-        // file has been uploaded successfull
-        // console.log("file is uploaded on cloudinary ", response.url);
-        fs.unlinkSync(localFilePath)
-        return response;
+const uploadOnCloudinary = async (localFilePath, upload_options) => {
+  try {
+    if (!localFilePath) return null;
+    //upload the file on cloudinary
+    const response = await cloudinary.uploader.upload(
+      localFilePath,
+      upload_options
+    );
+    if (response) fs.unlinkSync(localFilePath);
+    return response;
+  } catch (error) {
+    fs.unlinkSync(localFilePath);
+    console.log(error);
+    return null;
+  }
+};
 
-    } catch (error) {
-        fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
-        return null;
-    }
-}
+const deleteFromCloudinary = async (fileId) => {
+  try {
+    if (!fileId) return null;
+    //delete the file on cloudinary
+    const response = await cloudinary.uploader.destroy(fileId);
+    if (response) fs.unlinkSync(localFilePath);
+    return response;
+  } catch (error) {
+    return null;
+  }
+};
 
-
-
-export {uploadOnCloudinary}
-
-// mutler ko use karke -> local storage -> cloudinary -> cloud storage
+export { uploadOnCloudinary, deleteFromCloudinary };
